Update meeting weekly email picker
When changing the date for when we send weekly emails we want to ensure
that the meeting picker encompasses the dates that the meetings were
scheduled.

Also in case there was an issue sending the emails, this allows us to
manually trigger the emails without too much fuss in the console.